refactor: change currentOwner type from Stackholder to String in Product class
Updated the currentOwner field and associated methods to use a String instead of the Stackholder object. This simplifies the representation of the owner but may require adjustments to logic that relied on the Stackholder type.

BREAKING CHANGE: currentOwner is now a String, which could impact parts of the codebase expecting the previous Stackholder type.